requirements.txt
python-telegram-bot==20.7
yandexcloud==0.227.0
python-dotenv==1.0.0
requests 